DirSettings working
